% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouijaflow.R
\name{ouijaflow}
\alias{ouijaflow}
\title{Execute Wishbone}
\usage{
ouijaflow(expression, iter = 1000, num_cores = 1)
}
\arguments{
\item{expression}{Expression}

\item{iter}{Number of iterations}

\item{num_cores}{Number of cores allowed}
}
\description{
Execute Wishbone
}
