% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wishbone.R
\name{Wishbone}
\alias{Wishbone}
\title{Execute Wishbone}
\usage{
Wishbone(counts, start_cell_id, knn = 10, n_diffusion_components = 2,
  n_pca_components = 15, markers = "~", branch = TRUE, k = 15,
  num_waypoints = 50, normalize = TRUE, epsilon = 1, verbose = FALSE,
  num_cores = 1)
}
\arguments{
\item{counts}{Counts}

\item{start_cell_id}{ID of start cell}

\item{knn}{k nearest neighbours for diffusion}

\item{n_diffusion_components}{number of diffusion components}

\item{n_pca_components}{number of pca components}

\item{markers}{markers to use}

\item{branch}{whether or not to branch or linear}

\item{k}{k param}

\item{num_waypoints}{number of waypoints}

\item{normalize}{whether or not to normalize}

\item{epsilon}{epsilon param}

\item{verbose}{whether or not to print the wishbone output}

\item{num_cores}{number of cores to use}
}
\description{
Execute Wishbone
}
